Big refactoring of Views and Serializers, in the “event” application
