+++ event/views/events.py
@@ -1,0 +1,56 @@
+from rest_framework import status
+from rest_framework.response import Response
+from rest_framework.views import APIView
+
+from drf_yasg.utils import swagger_auto_schema
+
+from event.models import Event
+from event.serializers.events import EventSerializer
+
+from swagger_docs import SwaggerDocs
+
+
+class EventsListView(APIView):
+
+    @swagger_auto_schema(**SwaggerDocs.Event.post)
+    def post(self, request):
+        serializer = EventSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+class EventDetailView(APIView):
+    def get_object(self, pk):
+        return Event.objects.get(pk=pk)
+
+    @swagger_auto_schema(**SwaggerDocs.Event.get)
+    def get(self, request, pk):
+        event = self.get_object(pk)
+        serializer = EventSerializer(event)
+        return Response(serializer.data)
+
+    @swagger_auto_schema(**SwaggerDocs.Event.put)
+    def put(self, request, pk):
+        event = self.get_object(pk)
+        serializer = EventSerializer(event, data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    @swagger_auto_schema(**SwaggerDocs.Event.patch)
+    def patch(self, request, pk):
+        event = self.get_object(pk)
+        serializer = EventSerializer(event, data=request.data, partial=True)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    @swagger_auto_schema(**SwaggerDocs.Event.delete)
+    def delete(self, request, pk):
+        event = self.get_object(pk)
+        event.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
